feat: implement participant management endpoints in management API; add functionality for creating and editing virtual participants
