+++ services/management-api/apihandlers/study-management.go
@@ -60,6 +60,7 @@ func (h *HttpEndpoints) AddStudyManagementAPI(rg *gin.RouterGroup) {
 		h.addStudyActionEndpoints(studyGroup)
 		h.addStudyDataExporterEndpoints(studyGroup)
 		h.addStudyDataExplorerEndpoints(studyGroup)
+		h.addParticipantManagementEndpoints(studyGroup)
 	}
 }
 
@@ -854,17 +855,6 @@ func (h *HttpEndpoints) addStudyDataExplorerEndpoints(rg *gin.RouterGroup) {
 			},
 			nil,
 			h.getStudyParticipant,
-		))
-
-		participantsGroup.PUT("/:participantID", h.useAuthorisedHandler(
-			RequiredPermission{
-				ResourceType:        pc.RESOURCE_TYPE_STUDY,
-				ResourceKeys:        []string{pc.RESOURCE_KEY_STUDY_ALL},
-				ExtractResourceKeys: getStudyKeyFromParams,
-				Action:              pc.ACTION_EDIT_PARTICIPANT_STATES,
-			},
-			nil,
-			h.editStudyParticipant,
 		))
 	}
 
@@ -3533,40 +3523,6 @@ func (h *HttpEndpoints) getStudyParticipant(c *gin.Context) {
 	}
 
 	c.JSON(http.StatusOK, gin.H{"participant": participant})
-}
-
-func (h *HttpEndpoints) editStudyParticipant(c *gin.Context) {
-	token := c.MustGet("validatedToken").(*jwthandling.ManagementUserClaims)
-
-	studyKey := c.Param("studyKey")
-	participantID := c.Param("participantID")
-
-	slog.Info("updating participant", slog.String("participantID", participantID), slog.String("studyKey", studyKey), slog.String("userID", token.Subject), slog.String("instanceID", token.InstanceID))
-
-	var req studyTypes.Participant
-	if err := c.ShouldBindJSON(&req); err != nil {
-		slog.Error("failed to bind request", slog.String("error", err.Error()))
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid request"})
-		return
-	}
-
-	if req.ParticipantID != participantID {
-		slog.Error("participant ID in request does not match participant ID in path", slog.String("requestParticipantID", req.ParticipantID), slog.String("pathParticipantID", participantID))
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid request"})
-		return
-	}
-
-	updatedParticipant, err := h.studyDBConn.UpdateParticipantIfNotModified(token.InstanceID, studyKey, req)
-	if err != nil {
-		slog.Error("failed to update participant", slog.String("error", err.Error()))
-		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"message":     "participant updated",
-		"participant": updatedParticipant,
-	})
 }
 
 func (h *HttpEndpoints) getStudyReports(c *gin.Context) {
